pyproject.toml
[tool.poetry]
name = "basic-bot-net-lab"
version = "0.0.0"
description = "A containerized botnet managed by a virtual environment via Poerty"
authors = ["SubArtic15"]

[tool.poetry.dependencies]
python = "^3.10"
pexpect = "^4.8.0"
python-nmap = "^0.7.1"
termcolor = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
